Add rdf.common..RdfPartitionDefault basic unit tests
