+++ sansa-rdf-flink/src/main/scala/net/sansa_stack/rdf/flink/io/nquads/NQuadsReader.scala
@@ -161,7 +161,6 @@ object NQuadsReader {
           // filter out null values
           Iterators.filter(it, Predicates.notNull[Quad]())
         }
-      p
       new IteratorResourceClosing[Quad](it, input).asScala
     })
       .map(_.asTriple())
